+++ RELEASES.md
@@ -18,7 +18,8 @@ Release 0.11.0 on 2017-??-??
   by Martin Holmer]
 
 **New Features**
-- None
+- Add option to cap the amount of gross itemized deductions allowed as a decimal fraction of AGI
+  [[#1542](https://github.com/open-source-economics/Tax-Calculator/pull/1542) by Matt Jensen]
 
 **Bug Fixes**
 - None